requirements.txt
fastapi==0.112.1
PyYAML==6.0.2
Requests==2.32.3
selenium==4.23.1
uvicorn==0.30.6
